Refactoring: changes in the tests project
* removed unnecessary directives

* adjusted namespaces for easy navigation in the " Test Browser"
