+++ serverapp/tests/InfrastructureTests.Identity/UnitTests/IdentityTokenClaimServiceTest.cs
@@ -1,8 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.IdentityModel.Tokens.Jwt;
 
 using Xunit;
