+++ serverapp/tests/WebApi.Tests/Common/Utilities.cs
@@ -1,9 +1,9 @@
 ﻿using System;
 using System.Threading.Tasks;
-using Newtonsoft.Json;
-
 using System.Net.Http;
 using System.Text;
+
+using Newtonsoft.Json;
 
 using Infrastructure.Identity.Data;
 using Persistence.Context;
