+++ serverapp/tests/WebApi.Tests/IntrgrationTests/UserControllerEndpointsTests/SendConfirmationEmailEndpointTests.cs
@@ -5,11 +5,11 @@ using Xunit;
 
 using WebApi.Tests.Common;
 
-namespace WebApi.Tests.FunctionalTests.UserControllerTests
+namespace WebApi.Tests.IntegrationTests.UserControllEndpointsTests
 {
-    public class SendResetPasswordEmailEndpointTests:UserControllerEndpointTestsBase
+    public class SendConfirmationEmailEndpointTests:UserControllerEndpointTestsBase
     {
-        public SendResetPasswordEmailEndpointTests(ApiTestsFixture factory) : base(factory) { }
+        public SendConfirmationEmailEndpointTests(ApiTestsFixture factory) : base(factory) { }
 
         [Fact]
         public async Task ReturnsSuccessStatusCode()
@@ -17,7 +17,7 @@ namespace WebApi.Tests.FunctionalTests.UserControllerTests
             var client = _factory.GetAnonymousClient();
             var email = "test@test.com";
 
-            var response = await client.GetAsync($"api/user/{email}/resetpasswordemail");
+            var response = await client.GetAsync($"api/user/{email}/confirmationemail");
 
             response.EnsureSuccessStatusCode();
         }
@@ -28,7 +28,7 @@ namespace WebApi.Tests.FunctionalTests.UserControllerTests
             var client = _factory.GetAnonymousClient();
             var email = "invalid@test.com";
 
-            var response = await client.GetAsync($"api/user/{email}/resetpasswordemail");
+            var response = await client.GetAsync($"api/user/{email}/confirmationemail");
 
             Assert.Equal(HttpStatusCode.NotFound, response.StatusCode);
         }
