+++ serverapp/tests/WebApi.Tests/IntrgrationTests/UserControllerEndpointsTests/VerifyResetPasswordTokenEndpointTests.cs
@@ -9,7 +9,7 @@ using Xunit;
 
 using WebApi.Tests.Common;
 
-namespace WebApi.Tests.FunctionalTests.UserControllerTests
+namespace WebApi.Tests.IntegrationTests.UserControllEndpointsTests
 {
     public class VerifyResetPasswordTokenEndpointTests : UserControllerEndpointTestsBase
     {
@@ -50,7 +50,7 @@ namespace WebApi.Tests.FunctionalTests.UserControllerTests
         }
 
         [Fact]
-        public async Task ReturnsNotFoundStatusCodeWhenUserNotExis()
+        public async Task ReturnsNotFoundStatusCode_WhenUserNotExis()
         {
             var client = _factory.GetAnonymousClient();
             var userId = "invaliduserid";
